+++ plottr/gui/widgets.py
@@ -8,7 +8,8 @@ from typing import Union, List, Tuple, Optional, Type, Sequence, Dict, Any, Type
 
 from .tools import dictToTreeWidgetItems, dpiScalingFactor
 from plottr import QtGui, QtCore, Flowchart, QtWidgets, Signal, Slot
-from plottr.node import Node, linearFlowchart
+from plottr.node import Node, linearFlowchart, NodeWidget, updateOption
+from plottr.node.node import updateGuiQuietly, emitGuiUpdate
 from ..plot import PlotNode, PlotWidgetContainer, PlotWidget
 from .. import config_entry as getcfg
 
@@ -305,3 +306,51 @@ class Collapsible(QtWidgets.QWidget):
         else:
             self.widget.setVisible(False)
             self.btn.setText(self.collapsedTitle)
+
+
+class DimensionCombo(QtWidgets.QComboBox):
+    dimensionSelected = Signal(str)
+
+    def __init__(self, parent: Optional[QtWidgets.QWidget] = None,
+                 dimensionType: str = 'axes'):
+        super().__init__(parent)
+
+        self.node: Optional[Node] = None
+        self.dimensionType = dimensionType
+
+        self.clear()
+        self.entries = ['None']
+        for e in self.entries:
+            self.addItem(e)
+
+        self.currentTextChanged.connect(self.signalDimensionSelection)
+
+    def connectNode(self, node: Optional[Node] = None) -> None:
+        if node is None:
+            raise RuntimeError
+        self.node = node
+        if self.dimensionType == 'axes':
+            self.node.dataAxesChanged.connect(self.setDimensions)
+        else:
+            raise NotImplementedError('Only Axes supported ATM.')
+
+    @updateGuiQuietly
+    def setDimensions(self, dims: Sequence[str]) -> None:
+        self.clear()
+        allDims = self.entries + list(dims)
+        for d in allDims:
+            self.addItem(d)
+
+    @Slot(str)
+    @emitGuiUpdate('dimensionSelected')
+    def signalDimensionSelection(self, val: str) -> str:
+        return val
+
+
+class AxisSelector(FormLayoutWrapper):
+
+    def __init__(self, parent: Optional[QtWidgets.QWidget] = None):
+        super().__init__(
+            parent=parent,
+            elements=[('Axis', DimensionCombo(dimensionType='axes'))],
+        )
